Support dynamically computing embeddings for a catalog (#4)
Users can pass in a STAC catalog url with their search request.
The tool will fetch the collections from the mentioned catalog,
compute and cache the embeddings and search against the mentioned
catalog.

This would allow a stac-semantic-search instance to search against
any public STAC catalog and not just a fixed one